+++ src/app/generic/generic-create/generic-create.component.spec.ts
@@ -2,6 +2,11 @@ import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { GenericCreateComponent } from './generic-create.component';
 
+declare const describe: (description: string, specDefinitions: () => void) => void;
+declare const beforeEach: (action: ((done: unknown) => void) | (() => void) | (() => Promise<void>)) => void;
+declare const it: (expectation: string, assertion: () => void) => void;
+declare const expect: (actual: unknown) => { toBeTruthy: () => void };
+
 describe('GenericCreateComponent', () => {
   let component: GenericCreateComponent;
   let fixture: ComponentFixture<GenericCreateComponent>;