+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* 2026-07-26 - AB#175: Fix lint warnings in generic.
 * 2026-07-26 - AB#175: Fix lint warnings in common.
 * 2026-07-25 - AB#175: Remove deprecated baseUrl from tsconfig.json.
 * 2026-07-24 - Add project files.